[Decomp][Guild] Cohesion split: extract bank + rank management to GuildBank/GuildRank.cpp
Mirror of mangosthree #262. Pure cohesion split of Guild.cpp with no
behaviour change:

- GuildBank.cpp  - 37 bank methods: tab/content display, item
  store/withdraw/swap, money and rights tracking, bank event logging.
- GuildRank.cpp  - 7 rank methods: create/delete and name/rights.

Guild.h is unchanged. CMake file(GLOB Object/*.cpp) auto-compiles the
new translation units. WotLK has no SwitchRank (Cata-only), so GuildRank
carries 7 methods vs 8 in the three fork.
