+++ src/game/Object/GuildBank.cpp
@@ -1,0 +1,1682 @@
+/**
+ * MaNGOS is a full featured server for World of Warcraft, supporting
+ * the following clients: 1.12.x, 2.4.3, 3.3.5a, 4.3.4a and 5.4.8
+ *
+ * Copyright (C) 2005-2025 MaNGOS <https://www.getmangos.eu>
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+ *
+ * World of Warcraft, and all World of Warcraft or Warcraft art, images,
+ * and lore are copyrighted by Blizzard Entertainment, Inc.
+ */
+
+/**
+ * @file GuildBank.cpp
+ * @brief Cohesion split of Guild.cpp -- guild bank storage: tab/content
+ *        display, item store/withdraw/swap, money and rights tracking, and
+ *        bank event logging. Same `Guild` class; no behaviour change.
+ */
+
+#include "Guild.h"
+#include "GuildMgr.h"
+#include "Database/DatabaseEnv.h"
+#include "ObjectMgr.h"
+#include "ObjectAccessor.h"
+#include "Player.h"
+#include "Item.h"
+#include "World.h"
+#include "WorldSession.h"
+#include "WorldPacket.h"
+#include "Opcodes.h"
+#include "SocialMgr.h"
+#include "Chat.h"
+#include "Util.h"
+#include "Language.h"
+#include "Log.h"
+#ifdef ENABLE_ELUNA
+#include "LuaEngine.h"
+#endif /* ENABLE_ELUNA */
+
+// *************************************************
+// Guild Bank part
+// *************************************************
+// Bank content related
+void Guild::DisplayGuildBankContent(WorldSession* session, uint8 TabId)
+{
+    GuildBankTab const* tab = m_TabListMap[TabId];
+
+    if (!IsMemberHaveRights(session->GetPlayer()->GetGUIDLow(), TabId, GUILD_BANK_RIGHT_VIEW_TAB))
+    {
+        return;
+    }
+
+    WorldPacket data(SMSG_GUILD_BANK_LIST, 1200);
+
+    data << uint64(GetGuildBankMoney());
+    data << uint8(TabId);
+    // remaining slots for today
+    data << uint32(GetMemberSlotWithdrawRem(session->GetPlayer()->GetGUIDLow(), TabId));
+    data << uint8(0);                                       // Tell client that there's no tab info in this packet
+
+    data << uint8(GUILD_BANK_MAX_SLOTS);
+
+    for (int i = 0; i < GUILD_BANK_MAX_SLOTS; ++i)
+    {
+        AppendDisplayGuildBankSlot(data, tab, i);
+    }
+
+    session->SendPacket(&data);
+
+    DEBUG_LOG("WORLD: Sent (SMSG_GUILD_BANK_LIST)");
+}
+
+void Guild::DisplayGuildBankMoneyUpdate(WorldSession* session)
+{
+    WorldPacket data(SMSG_GUILD_BANK_LIST, 8 + 1 + 4 + 1 + 1);
+
+    data << uint64(GetGuildBankMoney());
+    data << uint8(0);                                       // TabId, default 0
+    data << uint32(GetMemberSlotWithdrawRem(session->GetPlayer()->GetGUIDLow(), 0));
+    data << uint8(0);                                       // Tell that there's no tab info in this packet
+    data << uint8(0);                                       // not send items
+    BroadcastPacket(&data);
+
+    DEBUG_LOG("WORLD: Sent (SMSG_GUILD_BANK_LIST)");
+}
+
+void Guild::DisplayGuildBankContentUpdate(uint8 TabId, int32 slot1, int32 slot2)
+{
+    GuildBankTab const* tab = m_TabListMap[TabId];
+
+    WorldPacket data(SMSG_GUILD_BANK_LIST, 1200);
+
+    data << uint64(GetGuildBankMoney());
+    data << uint8(TabId);
+
+    size_t rempos = data.wpos();
+    data << uint32(0);                                      // item withdraw amount, will be filled later
+    data << uint8(0);                                       // Tell client that there's no tab info in this packet
+
+    if (slot2 == -1)                                        // single item in slot1
+    {
+        data << uint8(1);                                   // item count
+
+        AppendDisplayGuildBankSlot(data, tab, slot1);
+    }
+    else                                                    // 2 items (in slot1 and slot2)
+    {
+        data << uint8(2);                                   // item count
+
+        if (slot1 > slot2)
+        {
+            std::swap(slot1, slot2);
+        }
+
+        AppendDisplayGuildBankSlot(data, tab, slot1);
+        AppendDisplayGuildBankSlot(data, tab, slot2);
+    }
+
+    for (MemberList::const_iterator itr = members.begin(); itr != members.end(); ++itr)
+    {
+        Player* player = sObjectAccessor.FindPlayer(ObjectGuid(HIGHGUID_PLAYER, itr->first));
+        if (!player)
+        {
+            continue;
+        }
+
+        if (!IsMemberHaveRights(itr->first, TabId, GUILD_BANK_RIGHT_VIEW_TAB))
+        {
+            continue;
+        }
+
+        data.put<uint32>(rempos, uint32(GetMemberSlotWithdrawRem(player->GetGUIDLow(), TabId)));
+
+        player->GetSession()->SendPacket(&data);
+    }
+
+    DEBUG_LOG("WORLD: Sent (SMSG_GUILD_BANK_LIST)");
+}
+
+void Guild::DisplayGuildBankContentUpdate(uint8 TabId, GuildItemPosCountVec const& slots)
+{
+    GuildBankTab const* tab = m_TabListMap[TabId];
+
+    WorldPacket data(SMSG_GUILD_BANK_LIST, 1200);
+
+    data << uint64(GetGuildBankMoney());
+    data << uint8(TabId);
+
+    size_t rempos = data.wpos();
+    data << uint32(0);                                      // item withdraw amount, will be filled later
+    data << uint8(0);                                       // Tell client that there's no tab info in this packet
+
+    data << uint8(slots.size());                            // updates count
+
+    for (GuildItemPosCountVec::const_iterator itr = slots.begin(); itr != slots.end(); ++itr)
+    {
+        AppendDisplayGuildBankSlot(data, tab, itr->Slot);
+    }
+
+    for (MemberList::const_iterator itr = members.begin(); itr != members.end(); ++itr)
+    {
+        Player* player = sObjectAccessor.FindPlayer(ObjectGuid(HIGHGUID_PLAYER, itr->first));
+        if (!player)
+        {
+            continue;
+        }
+
+        if (!IsMemberHaveRights(itr->first, TabId, GUILD_BANK_RIGHT_VIEW_TAB))
+        {
+            continue;
+        }
+
+        data.put<uint32>(rempos, uint32(GetMemberSlotWithdrawRem(player->GetGUIDLow(), TabId)));
+
+        player->GetSession()->SendPacket(&data);
+    }
+
+    DEBUG_LOG("WORLD: Sent (SMSG_GUILD_BANK_LIST)");
+}
+
+Item* Guild::GetItem(uint8 TabId, uint8 SlotId)
+{
+    if (TabId >= GetPurchasedTabs() || SlotId >= GUILD_BANK_MAX_SLOTS)
+    {
+        return NULL;
+    }
+    return m_TabListMap[TabId]->Slots[SlotId];
+}
+
+void Guild::DisplayGuildBankTabsInfo(WorldSession* session)
+{
+    WorldPacket data(SMSG_GUILD_BANK_LIST, 500);
+
+    data << uint64(GetGuildBankMoney());
+    data << uint8(0);                                       // TabInfo packet must be for TabId 0
+    data << uint32(GetMemberSlotWithdrawRem(session->GetPlayer()->GetGUIDLow(), 0));
+    data << uint8(1);                                       // Tell client that this packet includes tab info
+
+    data << uint8(GetPurchasedTabs());                      // here is the number of tabs
+
+    for (uint8 i = 0; i < GetPurchasedTabs(); ++i)
+    {
+        data << m_TabListMap[i]->Name.c_str();
+        data << m_TabListMap[i]->Icon.c_str();
+    }
+    data << uint8(0);                                       // Do not send tab content
+    session->SendPacket(&data);
+
+    DEBUG_LOG("WORLD: Sent (SMSG_GUILD_BANK_LIST)");
+}
+
+void Guild::CreateNewBankTab()
+{
+    if (GetPurchasedTabs() >= GUILD_BANK_MAX_TABS)
+    {
+        return;
+    }
+
+    uint32 tabId = GetPurchasedTabs();                      // next free id
+    m_TabListMap.push_back(new GuildBankTab);
+
+    CharacterDatabase.BeginTransaction();
+    CharacterDatabase.PExecute("DELETE FROM `guild_bank_tab` WHERE `guildid`='%u' AND `TabId`='%u'", m_Id, tabId);
+    CharacterDatabase.PExecute("INSERT INTO `guild_bank_tab` (`guildid`,`TabId`) VALUES ('%u','%u')", m_Id, tabId);
+    CharacterDatabase.CommitTransaction();
+}
+
+void Guild::SetGuildBankTabInfo(uint8 TabId, std::string Name, std::string Icon)
+{
+    if (m_TabListMap[TabId]->Name == Name && m_TabListMap[TabId]->Icon == Icon)
+    {
+        return;
+    }
+
+    m_TabListMap[TabId]->Name = Name;
+    m_TabListMap[TabId]->Icon = Icon;
+
+    CharacterDatabase.escape_string(Name);
+    CharacterDatabase.escape_string(Icon);
+    CharacterDatabase.PExecute("UPDATE `guild_bank_tab` SET `TabName`='%s',`TabIcon`='%s' WHERE `guildid`='%u' AND `TabId`='%u'", Name.c_str(), Icon.c_str(), m_Id, uint32(TabId));
+}
+
+uint32 Guild::GetBankRights(uint32 rankId, uint8 TabId) const
+{
+    if (rankId >= m_Ranks.size() || TabId >= GUILD_BANK_MAX_TABS)
+    {
+        return 0;
+    }
+
+    return m_Ranks[rankId].TabRight[TabId];
+}
+
+// This load should be called on startup only
+void Guild::LoadGuildBankFromDB()
+{
+    //                                                      0        1          2          3
+    QueryResult* result = CharacterDatabase.PQuery("SELECT `TabId`, `TabName`, `TabIcon`, `TabText` FROM `guild_bank_tab` WHERE `guildid`='%u' ORDER BY `TabId`", m_Id);
+    if (!result)
+    {
+        m_TabListMap.clear();
+        return;
+    }
+
+    do
+    {
+        Field* fields = result->Fetch();
+        uint8 tabId = fields[0].GetUInt8();
+        if (tabId >= GetPurchasedTabs())
+        {
+            sLog.outError("Table `guild_bank_tab` have not purchased tab %u for guild %u, skipped", tabId, m_Id);
+            continue;
+        }
+
+        GuildBankTab* NewTab = new GuildBankTab;
+
+        NewTab->Name = fields[1].GetCppString();
+        NewTab->Icon = fields[2].GetCppString();
+        NewTab->Text = fields[3].GetCppString();
+
+        m_TabListMap[tabId] = NewTab;
+    }
+    while (result->NextRow());
+
+    delete result;
+
+    // data needs to be at first place for Item::LoadFromDB
+    //                                        0     1     2      3       4          5
+    result = CharacterDatabase.PQuery("SELECT `data`, `text`, `TabId`, `SlotId`, `item_guid`, `item_entry` FROM `guild_bank_item` JOIN `item_instance` ON `item_guid` = `guid` WHERE `guildid`='%u' ORDER BY `TabId`", m_Id);
+    if (!result)
+    {
+        return;
+    }
+
+    do
+    {
+        Field* fields = result->Fetch();
+        uint8 TabId = fields[2].GetUInt8();
+        uint8 SlotId = fields[3].GetUInt8();
+        uint32 ItemGuid = fields[4].GetUInt32();
+        uint32 ItemEntry = fields[5].GetUInt32();
+
+        if (TabId >= GetPurchasedTabs())
+        {
+            sLog.outError("Guild::LoadGuildBankFromDB: Invalid tab for item (GUID: %u id: #%u) in guild bank, skipped.", ItemGuid, ItemEntry);
+            continue;
+        }
+
+        if (SlotId >= GUILD_BANK_MAX_SLOTS)
+        {
+            sLog.outError("Guild::LoadGuildBankFromDB: Invalid slot for item (GUID: %u id: #%u) in guild bank, skipped.", ItemGuid, ItemEntry);
+            continue;
+        }
+
+        ItemPrototype const* proto = ObjectMgr::GetItemPrototype(ItemEntry);
+
+        if (!proto)
+        {
+            sLog.outError("Guild::LoadGuildBankFromDB: Unknown item (GUID: %u id: #%u) in guild bank, skipped.", ItemGuid, ItemEntry);
+            continue;
+        }
+
+        Item* pItem = NewItemOrBag(proto);
+        if (!pItem->LoadFromDB(ItemGuid, fields))
+        {
+            CharacterDatabase.PExecute("DELETE FROM `guild_bank_item` WHERE `guildid`='%u' AND `TabId`='%u' AND `SlotId`='%u'", m_Id, uint32(TabId), uint32(SlotId));
+            sLog.outError("Item GUID %u not found in item_instance, deleting from Guild Bank!", ItemGuid);
+            delete pItem;
+            continue;
+        }
+
+        pItem->AddToWorld();
+        m_TabListMap[TabId]->Slots[SlotId] = pItem;
+    }
+    while (result->NextRow());
+
+    delete result;
+}
+
+void Guild::SendMoneyInfo(WorldSession* session, uint32 LowGuid)
+{
+    WorldPacket data(MSG_GUILD_BANK_MONEY_WITHDRAWN, 4);
+    data << uint32(GetMemberMoneyWithdrawRem(LowGuid));
+    session->SendPacket(&data);
+    DEBUG_LOG("WORLD: Sent MSG_GUILD_BANK_MONEY_WITHDRAWN");
+}
+
+bool Guild::MemberMoneyWithdraw(uint32 amount, uint32 LowGuid)
+{
+    uint32 MoneyWithDrawRight = GetMemberMoneyWithdrawRem(LowGuid);
+
+    if (MoneyWithDrawRight < amount || GetGuildBankMoney() < amount)
+    {
+        return false;
+    }
+
+    SetBankMoney(GetGuildBankMoney() - amount);
+
+    if (MoneyWithDrawRight < WITHDRAW_MONEY_UNLIMITED)
+    {
+        MemberList::iterator itr = members.find(LowGuid);
+        if (itr == members.end())
+        {
+            return false;
+        }
+        itr->second.BankRemMoney -= amount;
+        CharacterDatabase.PExecute("UPDATE `guild_member` SET `BankRemMoney`='%u' WHERE `guildid`='%u' AND `guid`='%u'",
+                                   itr->second.BankRemMoney, m_Id, LowGuid);
+    }
+
+#ifdef ENABLE_ELUNA
+    Player* player = sObjectMgr.GetPlayer(ObjectGuid(HIGHGUID_PLAYER, LowGuid));
+    if (Eluna* e = sWorld.GetEluna())
+    {
+        e->OnMemberWitdrawMoney(this, player, amount, false); // IsRepair not a part of Mangos, implement?
+    }
+#endif
+
+    return true;
+}
+
+void Guild::SetBankMoney(int64 money)
+{
+    if (money < 0)                                          // I don't know how this happens, it does!!
+    {
+        money = 0;
+    }
+    m_GuildBankMoney = money;
+
+    CharacterDatabase.PExecute("UPDATE `guild` SET `BankMoney`='" UI64FMTD "' WHERE `guildid`='%u'", money, m_Id);
+}
+
+bool Guild::MemberItemWithdraw(uint8 TabId, uint32 LowGuid)
+{
+    uint32 SlotsWithDrawRight = GetMemberSlotWithdrawRem(LowGuid, TabId);
+
+    if (SlotsWithDrawRight == 0)
+    {
+        return false;
+    }
+
+    if (SlotsWithDrawRight < WITHDRAW_SLOT_UNLIMITED)
+    {
+        MemberList::iterator itr = members.find(LowGuid);
+        if (itr == members.end())
+        {
+            return false;
+        }
+        --itr->second.BankRemSlotsTab[TabId];
+        CharacterDatabase.PExecute("UPDATE `guild_member` SET `BankRemSlotsTab%u`='%u' WHERE `guildid`='%u' AND `guid`='%u'",
+                                   uint32(TabId), itr->second.BankRemSlotsTab[TabId], m_Id, LowGuid);
+    }
+    return true;
+}
+
+bool Guild::IsMemberHaveRights(uint32 LowGuid, uint8 TabId, uint32 rights) const
+{
+    MemberList::const_iterator itr = members.find(LowGuid);
+    if (itr == members.end())
+    {
+        return false;
+    }
+
+    if (itr->second.RankId == GR_GUILDMASTER)
+    {
+        return true;
+    }
+
+    return (GetBankRights(itr->second.RankId, TabId) & rights) == rights;
+}
+
+uint32 Guild::GetMemberSlotWithdrawRem(uint32 LowGuid, uint8 TabId)
+{
+    MemberList::iterator itr = members.find(LowGuid);
+    if (itr == members.end())
+    {
+        return 0;
+    }
+
+    MemberSlot& member = itr->second;
+    if (member.RankId == GR_GUILDMASTER)
+    {
+        return WITHDRAW_SLOT_UNLIMITED;
+    }
+
+    if ((GetBankRights(member.RankId, TabId) & GUILD_BANK_RIGHT_VIEW_TAB) != GUILD_BANK_RIGHT_VIEW_TAB)
+    {
+        return 0;
+    }
+
+    uint32 curTime = uint32(time(NULL) / MINUTE);
+    if (curTime - member.BankResetTimeTab[TabId] >= 24 * HOUR / MINUTE)
+    {
+        member.BankResetTimeTab[TabId] = curTime;
+        member.BankRemSlotsTab[TabId] = GetBankSlotPerDay(member.RankId, TabId);
+        CharacterDatabase.PExecute("UPDATE `guild_member` SET `BankResetTimeTab%u`='%u', `BankRemSlotsTab%u`='%u' WHERE `guildid`='%u' AND `guid`='%u'",
+                                   uint32(TabId), member.BankResetTimeTab[TabId], uint32(TabId), member.BankRemSlotsTab[TabId], m_Id, LowGuid);
+    }
+    return member.BankRemSlotsTab[TabId];
+}
+
+uint32 Guild::GetMemberMoneyWithdrawRem(uint32 LowGuid)
+{
+    MemberList::iterator itr = members.find(LowGuid);
+    if (itr == members.end())
+    {
+        return 0;
+    }
+
+    MemberSlot& member = itr->second;
+    if (member.RankId == GR_GUILDMASTER)
+    {
+        return WITHDRAW_MONEY_UNLIMITED;
+    }
+
+    uint32 curTime = uint32(time(NULL) / MINUTE);           // minutes
+    // 24 hours
+    if (curTime > member.BankResetTimeMoney + 24 * HOUR / MINUTE)
+    {
+        member.BankResetTimeMoney = curTime;
+        member.BankRemMoney = GetBankMoneyPerDay(member.RankId);
+        CharacterDatabase.PExecute("UPDATE `guild_member` SET `BankResetTimeMoney`='%u', `BankRemMoney`='%u' WHERE `guildid`='%u' AND `guid`='%u'",
+                                   member.BankResetTimeMoney, member.BankRemMoney, m_Id, LowGuid);
+    }
+    return member.BankRemMoney;
+}
+
+void Guild::SetBankMoneyPerDay(uint32 rankId, uint32 money)
+{
+    if (rankId >= m_Ranks.size())
+    {
+        return;
+    }
+
+    if (rankId == GR_GUILDMASTER)
+    {
+        money = (uint32)WITHDRAW_MONEY_UNLIMITED;
+    }
+
+    m_Ranks[rankId].BankMoneyPerDay = money;
+
+    for (MemberList::iterator itr = members.begin(); itr != members.end(); ++itr)
+    {
+        MemberSlot& member = itr->second;
+        if (member.RankId == rankId)
+        {
+            member.BankResetTimeMoney = 0;
+        }
+    }
+
+    CharacterDatabase.PExecute("UPDATE `guild_rank` SET `BankMoneyPerDay`='%u' WHERE `rid`='%u' AND `guildid`='%u'", money, rankId, m_Id);
+    CharacterDatabase.PExecute("UPDATE `guild_member` SET `BankResetTimeMoney`='0' WHERE `guildid`='%u' AND `rank`='%u'", m_Id, rankId);
+}
+
+void Guild::SetBankRightsAndSlots(uint32 rankId, uint8 TabId, uint32 right, uint32 nbSlots, bool db)
+{
+    if (rankId >= m_Ranks.size() || TabId >= GetPurchasedTabs())
+    {
+        // TODO remove next line, It is there just to repair existing bug in deleting guild rank
+        CharacterDatabase.PExecute("DELETE FROM `guild_bank_right` WHERE `guildid`='%u' AND `rid`='%u' AND `TabId`='%u'", m_Id, rankId, TabId);
+        return;
+    }
+
+    if (rankId == GR_GUILDMASTER)
+    {
+        nbSlots = WITHDRAW_SLOT_UNLIMITED;
+        right = GUILD_BANK_RIGHT_FULL;
+    }
+
+    m_Ranks[rankId].TabSlotPerDay[TabId] = nbSlots;
+    m_Ranks[rankId].TabRight[TabId] = right;
+
+    if (db)
+    {
+        for (MemberList::iterator itr = members.begin(); itr != members.end(); ++itr)
+        {
+            if (itr->second.RankId == rankId)
+                for (int i = 0; i < GUILD_BANK_MAX_TABS; ++i)
+                {
+                    itr->second.BankResetTimeTab[i] = 0;
+                }
+        }
+
+        CharacterDatabase.PExecute("DELETE FROM `guild_bank_right` WHERE `guildid`='%u' AND `TabId`='%u' AND `rid`='%u'", m_Id, uint32(TabId), rankId);
+        CharacterDatabase.PExecute("INSERT INTO `guild_bank_right` (`guildid`,`TabId`,`rid`,`gbright`,`SlotPerDay`) VALUES "
+                                   "('%u','%u','%u','%u','%u')", m_Id, uint32(TabId), rankId, m_Ranks[rankId].TabRight[TabId], m_Ranks[rankId].TabSlotPerDay[TabId]);
+        CharacterDatabase.PExecute("UPDATE `guild_member` SET `BankResetTimeTab%u`='0' WHERE `guildid`='%u' AND `rank`='%u'", uint32(TabId), m_Id, rankId);
+    }
+}
+
+uint32 Guild::GetBankMoneyPerDay(uint32 rankId)
+{
+    if (rankId >= m_Ranks.size())
+    {
+        return 0;
+    }
+
+    if (rankId == GR_GUILDMASTER)
+    {
+        return (uint32)WITHDRAW_MONEY_UNLIMITED;
+    }
+    return m_Ranks[rankId].BankMoneyPerDay;
+}
+
+uint32 Guild::GetBankSlotPerDay(uint32 rankId, uint8 TabId)
+{
+    if (rankId >= m_Ranks.size() || TabId >= GUILD_BANK_MAX_TABS)
+    {
+        return 0;
+    }
+
+    if (rankId == GR_GUILDMASTER)
+    {
+        return WITHDRAW_SLOT_UNLIMITED;
+    }
+    return m_Ranks[rankId].TabSlotPerDay[TabId];
+}
+
+bool Guild::LoadBankRightsFromDB(QueryResult* guildBankTabRightsResult)
+{
+    if (!guildBankTabRightsResult)
+    {
+        return true;
+    }
+
+    do
+    {
+        Field* fields      = guildBankTabRightsResult->Fetch();
+        // prevent crash when all rights in result are already processed
+        if (!fields)
+        {
+            break;
+        }
+        uint32 guildId     = fields[0].GetUInt32();
+        if (guildId < m_Id)
+        {
+            // there is in table guild_bank_right record which doesn't have guildid in guild table, report error
+            sLog.outErrorDb("Guild %u does not exist but it has a record in guild_bank_right table, deleting it!", guildId);
+            CharacterDatabase.PExecute("DELETE FROM `guild_bank_right` WHERE `guildid` = '%u'", guildId);
+            continue;
+        }
+
+        if (guildId > m_Id)
+            // we loaded all ranks for this guild bank already, break cycle
+            break;
+        uint8 TabId        = fields[1].GetUInt8();
+        uint32 rankId      = fields[2].GetUInt32();
+        uint16 right       = fields[3].GetUInt16();
+        uint16 SlotPerDay  = fields[4].GetUInt16();
+
+        SetBankRightsAndSlots(rankId, TabId, right, SlotPerDay, false);
+    }
+    while (guildBankTabRightsResult->NextRow());
+
+    return true;
+}
+
+void Guild::LoadGuildBankEventLogFromDB()
+{
+    // Money log is in TabId = GUILD_BANK_MONEY_LOGS_TAB
+
+    // uint32 configCount = sWorld.getConfig(CONFIG_UINT32_GUILD_BANK_EVENT_LOG_COUNT);
+    // cycle through all purchased guild bank item tabs
+    for (uint32 tabId = 0; tabId < uint32(GetPurchasedTabs()); ++tabId)
+    {
+        //                                                      0          1            2             3              4                 5            6
+        QueryResult* result = CharacterDatabase.PQuery("SELECT `LogGuid`, `EventType`, `PlayerGuid`, `ItemOrMoney`, `ItemStackCount`, `DestTabId`, `TimeStamp` FROM `guild_bank_eventlog` WHERE `guildid`='%u' AND `TabId`='%u' ORDER BY `TimeStamp` DESC,`LogGuid` DESC LIMIT %u", m_Id, tabId, GUILD_BANK_MAX_LOGS);
+        if (!result)
+        {
+            continue;
+        }
+
+        bool isNextLogGuidSet = false;
+        do
+        {
+            Field* fields = result->Fetch();
+
+            GuildBankEventLogEntry NewEvent;
+            NewEvent.EventType = fields[1].GetUInt8();
+            NewEvent.PlayerGuid = fields[2].GetUInt32();
+            NewEvent.ItemOrMoney = fields[3].GetUInt32();
+            NewEvent.ItemStackCount = fields[4].GetUInt8();
+            NewEvent.DestTabId = fields[5].GetUInt8();
+            NewEvent.TimeStamp = fields[6].GetUInt64();
+
+            // if newEvent is moneyEvent, move it to moneyEventTab in DB and report error
+            if (NewEvent.isMoneyEvent())
+            {
+                uint32 logGuid = fields[0].GetUInt32();
+                CharacterDatabase.PExecute("UPDATE `guild_bank_eventlog` SET `TabId`='%u' WHERE `guildid`='%u' AND `TabId`='%u' AND `LogGuid`='%u'", GUILD_BANK_MONEY_LOGS_TAB, m_Id, tabId, logGuid);
+                sLog.outError("GuildBankEventLog ERROR: MoneyEvent LogGuid %u for Guild %u had incorrectly set its TabId to %u, correcting it to %u TabId", logGuid, m_Id, tabId, GUILD_BANK_MONEY_LOGS_TAB);
+                continue;
+            }
+            else
+                // add event to list
+                // events are ordered from oldest (in beginning) to latest (in the end)
+                m_GuildBankEventLog_Item[tabId].push_front(NewEvent);
+
+            if (!isNextLogGuidSet)
+            {
+                m_GuildBankEventLogNextGuid_Item[tabId] = fields[0].GetUInt32();
+                // we don't have to do m_GuildBankEventLogNextGuid_Item[tabId] %= configCount; - it will be done when creating new record
+                isNextLogGuidSet = true;
+            }
+        }
+        while (result->NextRow());
+        delete result;
+    }
+
+    // special handle for guild bank money log
+    //                                                      0          1            2             3              4                 5            6
+    QueryResult* result = CharacterDatabase.PQuery("SELECT `LogGuid`, `EventType`, `PlayerGuid`, `ItemOrMoney`, `ItemStackCount`, `DestTabId`, `TimeStamp` FROM `guild_bank_eventlog` WHERE `guildid`='%u' AND `TabId`='%u' ORDER BY `TimeStamp` DESC,`LogGuid` DESC LIMIT %u", m_Id, GUILD_BANK_MONEY_LOGS_TAB, GUILD_BANK_MAX_LOGS);
+    if (!result)
+    {
+        return;
+    }
+
+    bool isNextMoneyLogGuidSet = false;
+    do
+    {
+        Field* fields = result->Fetch();
+        if (!isNextMoneyLogGuidSet)
+        {
+            m_GuildBankEventLogNextGuid_Money = fields[0].GetUInt32();
+            // we don't have to do m_GuildBankEventLogNextGuid_Money %= configCount; - it will be done when creating new record
+            isNextMoneyLogGuidSet = true;
+        }
+        GuildBankEventLogEntry NewEvent;
+
+        NewEvent.EventType = fields[1].GetUInt8();
+        NewEvent.PlayerGuid = fields[2].GetUInt32();
+        NewEvent.ItemOrMoney = fields[3].GetUInt32();
+        NewEvent.ItemStackCount = fields[4].GetUInt8();
+        NewEvent.DestTabId = fields[5].GetUInt8();
+        NewEvent.TimeStamp = fields[6].GetUInt64();
+
+        // if newEvent is not moneyEvent, then report error
+        if (!NewEvent.isMoneyEvent())
+            sLog.outError("GuildBankEventLog ERROR: MoneyEvent LogGuid %u for Guild %u is not MoneyEvent - ignoring...", fields[0].GetUInt32(), m_Id);
+        else
+            // add event to list
+            // events are ordered from oldest (in beginning) to latest (in the end)
+            m_GuildBankEventLog_Money.push_front(NewEvent);
+    }
+    while (result->NextRow());
+    delete result;
+}
+
+void Guild::DisplayGuildBankLogs(WorldSession* session, uint8 TabId)
+{
+    if (TabId > GUILD_BANK_MAX_TABS)
+    {
+        return;
+    }
+
+    if (TabId == GUILD_BANK_MAX_TABS)
+    {
+        // Here we display money logs
+        WorldPacket data(MSG_GUILD_BANK_LOG_QUERY, m_GuildBankEventLog_Money.size()*(4 * 4 + 1) + 1 + 1);
+        data << uint8(TabId);                               // Here GUILD_BANK_MAX_TABS
+        data << uint8(m_GuildBankEventLog_Money.size());    // number of log entries
+        for (GuildBankEventLog::const_iterator itr = m_GuildBankEventLog_Money.begin(); itr != m_GuildBankEventLog_Money.end(); ++itr)
+        {
+            data << uint8(itr->EventType);
+            data << ObjectGuid(HIGHGUID_PLAYER, itr->PlayerGuid);
+            if (itr->EventType == GUILD_BANK_LOG_DEPOSIT_MONEY ||
+                    itr->EventType == GUILD_BANK_LOG_WITHDRAW_MONEY ||
+                    itr->EventType == GUILD_BANK_LOG_REPAIR_MONEY ||
+                    itr->EventType == GUILD_BANK_LOG_UNK1 ||
+                    itr->EventType == GUILD_BANK_LOG_UNK2)
+            {
+                data << uint32(itr->ItemOrMoney);
+            }
+            else
+            {
+                data << uint32(itr->ItemOrMoney);
+                data << uint32(itr->ItemStackCount);
+                if (itr->EventType == GUILD_BANK_LOG_MOVE_ITEM || itr->EventType == GUILD_BANK_LOG_MOVE_ITEM2)
+                {
+                    data << uint8(itr->DestTabId);          // moved tab
+                }
+            }
+            data << uint32(time(NULL) - itr->TimeStamp);
+        }
+        session->SendPacket(&data);
+    }
+    else
+    {
+        // here we display current tab logs
+        WorldPacket data(MSG_GUILD_BANK_LOG_QUERY, m_GuildBankEventLog_Item[TabId].size()*(4 * 4 + 1 + 1) + 1 + 1);
+        data << uint8(TabId);                               // Here a real Tab Id
+        // number of log entries
+        data << uint8(m_GuildBankEventLog_Item[TabId].size());
+        for (GuildBankEventLog::const_iterator itr = m_GuildBankEventLog_Item[TabId].begin(); itr != m_GuildBankEventLog_Item[TabId].end(); ++itr)
+        {
+            data << uint8(itr->EventType);
+            data << ObjectGuid(HIGHGUID_PLAYER, itr->PlayerGuid);
+            if (itr->EventType == GUILD_BANK_LOG_DEPOSIT_MONEY ||
+                    itr->EventType == GUILD_BANK_LOG_WITHDRAW_MONEY ||
+                    itr->EventType == GUILD_BANK_LOG_REPAIR_MONEY ||
+                    itr->EventType == GUILD_BANK_LOG_UNK1 ||
+                    itr->EventType == GUILD_BANK_LOG_UNK2)
+            {
+                data << uint32(itr->ItemOrMoney);
+            }
+            else
+            {
+                data << uint32(itr->ItemOrMoney);
+                data << uint32(itr->ItemStackCount);
+                if (itr->EventType == GUILD_BANK_LOG_MOVE_ITEM || itr->EventType == GUILD_BANK_LOG_MOVE_ITEM2)
+                {
+                    data << uint8(itr->DestTabId);          // moved tab
+                }
+            }
+            data << uint32(time(NULL) - itr->TimeStamp);
+        }
+        session->SendPacket(&data);
+    }
+    DEBUG_LOG("WORLD: Sent (MSG_GUILD_BANK_LOG_QUERY)");
+}
+
+void Guild::LogBankEvent(uint8 EventType, uint8 TabId, uint32 PlayerGuidLow, uint32 ItemOrMoney, uint8 ItemStackCount, uint8 DestTabId)
+{
+    // create Event
+    GuildBankEventLogEntry NewEvent;
+    NewEvent.EventType = EventType;
+    NewEvent.PlayerGuid = PlayerGuidLow;
+    NewEvent.ItemOrMoney = ItemOrMoney;
+    NewEvent.ItemStackCount = ItemStackCount;
+    NewEvent.DestTabId = DestTabId;
+    NewEvent.TimeStamp = uint32(time(NULL));
+
+    // add new event to the end of event list
+    uint32 currentTabId = TabId;
+    uint32 currentLogGuid = 0;
+    if (NewEvent.isMoneyEvent())
+    {
+        m_GuildBankEventLogNextGuid_Money = (m_GuildBankEventLogNextGuid_Money + 1) % sWorld.getConfig(CONFIG_UINT32_GUILD_BANK_EVENT_LOG_COUNT);
+        currentLogGuid = m_GuildBankEventLogNextGuid_Money;
+        currentTabId = GUILD_BANK_MONEY_LOGS_TAB;
+        if (m_GuildBankEventLog_Money.size() >= GUILD_BANK_MAX_LOGS)
+        {
+            m_GuildBankEventLog_Money.pop_front();
+        }
+
+        m_GuildBankEventLog_Money.push_back(NewEvent);
+    }
+    else
+    {
+        m_GuildBankEventLogNextGuid_Item[TabId] = ((m_GuildBankEventLogNextGuid_Item[TabId]) + 1) % sWorld.getConfig(CONFIG_UINT32_GUILD_BANK_EVENT_LOG_COUNT);
+        currentLogGuid = m_GuildBankEventLogNextGuid_Item[TabId];
+        if (m_GuildBankEventLog_Item[TabId].size() >= GUILD_BANK_MAX_LOGS)
+        {
+            m_GuildBankEventLog_Item[TabId].pop_front();
+        }
+
+        m_GuildBankEventLog_Item[TabId].push_back(NewEvent);
+    }
+
+#ifdef ENABLE_ELUNA
+    if (Eluna* e = sWorld.GetEluna())
+    {
+        e->OnBankEvent(this, EventType, TabId, PlayerGuidLow, ItemOrMoney, ItemStackCount, DestTabId);
+    }
+#endif
+
+    // save event to database
+    CharacterDatabase.PExecute("DELETE FROM `guild_bank_eventlog` WHERE `guildid`='%u' AND `LogGuid`='%u' AND `TabId`='%u'", m_Id, currentLogGuid, currentTabId);
+
+    CharacterDatabase.PExecute("INSERT INTO `guild_bank_eventlog` (`guildid`,`LogGuid`,`TabId`,`EventType`,`PlayerGuid`,`ItemOrMoney`,`ItemStackCount`,`DestTabId`,`TimeStamp`) VALUES ('%u','%u','%u','%u','%u','%u','%u','%u','" UI64FMTD "')",
+                               m_Id, currentLogGuid, currentTabId, uint32(NewEvent.EventType), NewEvent.PlayerGuid, NewEvent.ItemOrMoney, uint32(NewEvent.ItemStackCount), uint32(NewEvent.DestTabId), NewEvent.TimeStamp);
+}
+
+bool Guild::AddGBankItemToDB(uint32 GuildId, uint32 BankTab , uint32 BankTabSlot , uint32 GUIDLow, uint32 Entry)
+{
+    CharacterDatabase.PExecute("DELETE FROM `guild_bank_item` WHERE `guildid` = '%u' AND `TabId` = '%u'AND `SlotId` = '%u'", GuildId, BankTab, BankTabSlot);
+    CharacterDatabase.PExecute("INSERT INTO `guild_bank_item` (`guildid`,`TabId`,`SlotId`,`item_guid`,`item_entry`) "
+                               "VALUES ('%u', '%u', '%u', '%u', '%u')", GuildId, BankTab, BankTabSlot, GUIDLow, Entry);
+    return true;
+}
+
+void Guild::AppendDisplayGuildBankSlot(WorldPacket& data, GuildBankTab const* tab, int slot)
+{
+    Item* pItem = tab->Slots[slot];
+    uint32 entry = pItem ? pItem->GetEntry() : 0;
+
+    data << uint8(slot);
+    data << uint32(entry);
+    if (entry)
+    {
+        data << uint32(0);                                  // 3.3.0 (0x8000, 0x8020)
+        data << uint32(pItem->GetItemRandomPropertyId());   // random item property id + 8
+
+        if (pItem->GetItemRandomPropertyId())
+        {
+            data << uint32(pItem->GetItemSuffixFactor());   // SuffixFactor + 4
+        }
+
+        data << uint32(pItem->GetCount());                  // +12 ITEM_FIELD_STACK_COUNT
+        data << uint32(0);                                  // +16 Unknown value
+        data << uint8(0);                                   // +20
+
+        uint8 enchCount = 0;
+        size_t enchCountPos = data.wpos();
+
+        data << uint8(enchCount);                           // number of enchantments
+        for (uint32 i = PERM_ENCHANTMENT_SLOT; i < MAX_ENCHANTMENT_SLOT; ++i)
+        {
+            if (uint32 enchId = pItem->GetEnchantmentId(EnchantmentSlot(i)))
+            {
+                data << uint8(i);
+                data << uint32(enchId);
+                ++enchCount;
+            }
+        }
+        data.put<uint8>(enchCountPos, enchCount);
+    }
+}
+
+Item* Guild::StoreItem(uint8 tabId, GuildItemPosCountVec const& dest, Item* pItem)
+{
+    if (!pItem)
+    {
+        return NULL;
+    }
+
+    Item* lastItem = pItem;
+
+    for (GuildItemPosCountVec::const_iterator itr = dest.begin(); itr != dest.end();)
+    {
+        uint8 slot = itr->Slot;
+        uint32 count = itr->Count;
+
+        ++itr;
+
+        if (itr == dest.end())
+        {
+            lastItem = _StoreItem(tabId, slot, pItem, count, false);
+            break;
+        }
+
+        lastItem = _StoreItem(tabId, slot, pItem, count, true);
+    }
+
+    return lastItem;
+}
+
+// Return stored item (if stored to stack, it can diff. from pItem). And pItem ca be deleted in this case.
+Item* Guild::_StoreItem(uint8 tab, uint8 slot, Item* pItem, uint32 count, bool clone)
+{
+    if (!pItem)
+    {
+        return NULL;
+    }
+
+    DEBUG_LOG("GUILD STORAGE: StoreItem tab = %u, slot = %u, item = %u, count = %u", tab, slot, pItem->GetEntry(), count);
+
+    Item* pItem2 = m_TabListMap[tab]->Slots[slot];
+
+    if (!pItem2)
+    {
+        if (clone)
+        {
+            pItem = pItem->CloneItem(count);
+        }
+        else
+        {
+            pItem->SetCount(count);
+        }
+
+        if (!pItem)
+        {
+            return NULL;
+        }
+
+        m_TabListMap[tab]->Slots[slot] = pItem;
+
+        pItem->SetGuidValue(ITEM_FIELD_CONTAINED, ObjectGuid());
+        pItem->SetGuidValue(ITEM_FIELD_OWNER, ObjectGuid());
+        AddGBankItemToDB(GetId(), tab, slot, pItem->GetGUIDLow(), pItem->GetEntry());
+        pItem->FSetState(ITEM_NEW);
+        pItem->SaveToDB();                                  // not in inventory and can be save standalone
+
+        return pItem;
+    }
+    else
+    {
+        pItem2->SetCount(pItem2->GetCount() + count);
+        pItem2->FSetState(ITEM_CHANGED);
+        pItem2->SaveToDB();                                 // not in inventory and can be save standalone
+
+        if (!clone)
+        {
+            pItem->RemoveFromWorld();
+            pItem->DeleteFromDB();
+            delete pItem;
+        }
+
+        return pItem2;
+    }
+}
+
+void Guild::RemoveItem(uint8 tab, uint8 slot)
+{
+    m_TabListMap[tab]->Slots[slot] = NULL;
+    CharacterDatabase.PExecute("DELETE FROM `guild_bank_item` WHERE `guildid`='%u' AND `TabId`='%u' AND `SlotId`='%u'",
+                               GetId(), uint32(tab), uint32(slot));
+}
+
+InventoryResult Guild::_CanStoreItem_InSpecificSlot(uint8 tab, uint8 slot, GuildItemPosCountVec& dest, uint32& count, bool swap, Item* pSrcItem) const
+{
+    Item* pItem2 = m_TabListMap[tab]->Slots[slot];
+
+    // ignore move item (this slot will be empty at move)
+    if (pItem2 == pSrcItem)
+    {
+        pItem2 = NULL;
+    }
+
+    uint32 need_space;
+
+    // empty specific slot - check item fit to slot
+    if (!pItem2 || swap)
+    {
+        // non empty stack with space
+        need_space = pSrcItem->GetMaxStackCount();
+    }
+    // non empty slot, check item type
+    else
+    {
+        // check item type
+        if (pItem2->GetEntry() != pSrcItem->GetEntry())
+        {
+            return EQUIP_ERR_ITEM_CANT_STACK;
+        }
+
+        // check free space
+        if (pItem2->GetCount() >= pSrcItem->GetMaxStackCount())
+        {
+            return EQUIP_ERR_ITEM_CANT_STACK;
+        }
+
+        need_space = pSrcItem->GetMaxStackCount() - pItem2->GetCount();
+    }
+
+    if (need_space > count)
+    {
+        need_space = count;
+    }
+
+    GuildItemPosCount newPosition = GuildItemPosCount(slot, need_space);
+    if (!newPosition.isContainedIn(dest))
+    {
+        dest.push_back(newPosition);
+        count -= need_space;
+    }
+
+    return EQUIP_ERR_OK;
+}
+
+InventoryResult Guild::_CanStoreItem_InTab(uint8 tab, GuildItemPosCountVec& dest, uint32& count, bool merge, Item* pSrcItem, uint8 skip_slot) const
+{
+    for (uint32 j = 0; j < GUILD_BANK_MAX_SLOTS; ++j)
+    {
+        // skip specific slot already processed in first called _CanStoreItem_InSpecificSlot
+        if (j == skip_slot)
+        {
+            continue;
+        }
+
+        Item* pItem2 = m_TabListMap[tab]->Slots[j];
+
+        // ignore move item (this slot will be empty at move)
+        if (pItem2 == pSrcItem)
+        {
+            pItem2 = NULL;
+        }
+
+        // if merge skip empty, if !merge skip non-empty
+        if ((pItem2 != NULL) != merge)
+        {
+            continue;
+        }
+
+        if (pItem2)
+        {
+            if (pItem2->GetEntry() == pSrcItem->GetEntry() && pItem2->GetCount() < pSrcItem->GetMaxStackCount())
+            {
+                uint32 need_space = pSrcItem->GetMaxStackCount() - pItem2->GetCount();
+                if (need_space > count)
+                {
+                    need_space = count;
+                }
+
+                GuildItemPosCount newPosition = GuildItemPosCount(j, need_space);
+                if (!newPosition.isContainedIn(dest))
+                {
+                    dest.push_back(newPosition);
+                    count -= need_space;
+
+                    if (count == 0)
+                    {
+                        return EQUIP_ERR_OK;
+                    }
+                }
+            }
+        }
+        else
+        {
+            uint32 need_space = pSrcItem->GetMaxStackCount();
+            if (need_space > count)
+            {
+                need_space = count;
+            }
+
+            GuildItemPosCount newPosition = GuildItemPosCount(j, need_space);
+            if (!newPosition.isContainedIn(dest))
+            {
+                dest.push_back(newPosition);
+                count -= need_space;
+
+                if (count == 0)
+                {
+                    return EQUIP_ERR_OK;
+                }
+            }
+        }
+    }
+    return EQUIP_ERR_OK;
+}
+
+InventoryResult Guild::CanStoreItem(uint8 tab, uint8 slot, GuildItemPosCountVec& dest, uint32 count, Item* pItem, bool swap) const
+{
+    DEBUG_LOG("GUILD STORAGE: CanStoreItem tab = %u, slot = %u, item = %u, count = %u", tab, slot, pItem->GetEntry(), count);
+
+    if (count > pItem->GetCount())
+    {
+        return EQUIP_ERR_COULDNT_SPLIT_ITEMS;
+    }
+
+    if (pItem->IsSoulBound())
+    {
+        return EQUIP_ERR_CANT_DROP_SOULBOUND;
+    }
+
+    // in specific slot
+    if (slot != NULL_SLOT)
+    {
+        InventoryResult res = _CanStoreItem_InSpecificSlot(tab, slot, dest, count, swap, pItem);
+        if (res != EQUIP_ERR_OK)
+        {
+            return res;
+        }
+
+        if (count == 0)
+        {
+            return EQUIP_ERR_OK;
+        }
+    }
+
+    // not specific slot or have space for partly store only in specific slot
+
+    // search stack in tab for merge to
+    if (pItem->GetMaxStackCount() > 1)
+    {
+        InventoryResult res = _CanStoreItem_InTab(tab, dest, count, true, pItem, slot);
+        if (res != EQUIP_ERR_OK)
+        {
+            return res;
+        }
+
+        if (count == 0)
+        {
+            return EQUIP_ERR_OK;
+        }
+    }
+
+    // search free slot in bag for place to
+    InventoryResult res = _CanStoreItem_InTab(tab, dest, count, false, pItem, slot);
+    if (res != EQUIP_ERR_OK)
+    {
+        return res;
+    }
+
+    if (count == 0)
+    {
+        return EQUIP_ERR_OK;
+    }
+
+    return EQUIP_ERR_BANK_FULL;
+}
+
+void Guild::SetGuildBankTabText(uint8 TabId, std::string text)
+{
+    if (TabId >= GetPurchasedTabs())
+    {
+        return;
+    }
+
+    if (!m_TabListMap[TabId])
+    {
+        return;
+    }
+
+    if (m_TabListMap[TabId]->Text == text)
+    {
+        return;
+    }
+
+    utf8truncate(text, 500);                                // DB and client size limitation
+
+    m_TabListMap[TabId]->Text = text;
+
+    CharacterDatabase.escape_string(text);
+    CharacterDatabase.PExecute("UPDATE `guild_bank_tab` SET `TabText`='%s' WHERE `guildid`='%u' AND `TabId`='%u'", text.c_str(), m_Id, uint32(TabId));
+
+    // announce
+    SendGuildBankTabText(NULL, TabId);
+}
+
+void Guild::SendGuildBankTabText(WorldSession* session, uint8 TabId)
+{
+    GuildBankTab const* tab = m_TabListMap[TabId];
+
+    WorldPacket data(MSG_QUERY_GUILD_BANK_TEXT, 1 + tab->Text.size() + 1);
+    data << uint8(TabId);
+    data << tab->Text;
+
+    if (session)
+    {
+        session->SendPacket(&data);
+    }
+    else
+    {
+        BroadcastPacket(&data);
+    }
+}
+
+void Guild::SwapItems(Player* pl, uint8 BankTab, uint8 BankTabSlot, uint8 BankTabDst, uint8 BankTabSlotDst, uint32 SplitedAmount)
+{
+    // empty operation
+    if (BankTab == BankTabDst && BankTabSlot == BankTabSlotDst)
+    {
+        return;
+    }
+
+    Item* pItemSrc = GetItem(BankTab, BankTabSlot);
+    if (!pItemSrc)                                      // may prevent crash
+    {
+        return;
+    }
+
+    if (SplitedAmount > pItemSrc->GetCount())
+    {
+        return;                                         // cheating?
+    }
+    else if (SplitedAmount == pItemSrc->GetCount())
+    {
+        SplitedAmount = 0;                              // no split
+    }
+
+    Item* pItemDst = GetItem(BankTabDst, BankTabSlotDst);
+
+    if (BankTab != BankTabDst)
+    {
+        // check dest pos rights (if different tabs)
+        if (!IsMemberHaveRights(pl->GetGUIDLow(), BankTabDst, GUILD_BANK_RIGHT_DEPOSIT_ITEM))
+        {
+            return;
+        }
+
+        // check source pos rights (if different tabs)
+        uint32 remRight = GetMemberSlotWithdrawRem(pl->GetGUIDLow(), BankTab);
+        if (remRight <= 0)
+        {
+            return;
+        }
+    }
+
+    if (SplitedAmount)
+    {
+        // Bank -> Bank item split (in empty or non empty slot
+        GuildItemPosCountVec dest;
+        InventoryResult msg = CanStoreItem(BankTabDst, BankTabSlotDst, dest, SplitedAmount, pItemSrc, false);
+        if (msg != EQUIP_ERR_OK)
+        {
+            pl->SendEquipError(msg, pItemSrc, NULL);
+            return;
+        }
+
+        Item* pNewItem = pItemSrc->CloneItem(SplitedAmount);
+        if (!pNewItem)
+        {
+            pl->SendEquipError(EQUIP_ERR_ITEM_NOT_FOUND, pItemSrc, NULL);
+            return;
+        }
+
+        CharacterDatabase.BeginTransaction();
+
+        if (BankTab != BankTabDst)
+        {
+            LogBankEvent(GUILD_BANK_LOG_MOVE_ITEM, BankTab, pl->GetGUIDLow(), pItemSrc->GetEntry(), SplitedAmount, BankTabDst);
+        }
+
+        pl->ItemRemovedQuestCheck(pItemSrc->GetEntry(), SplitedAmount);
+        pItemSrc->SetCount(pItemSrc->GetCount() - SplitedAmount);
+        pItemSrc->FSetState(ITEM_CHANGED);
+        pItemSrc->SaveToDB();                               // not in inventory and can be save standalone
+        StoreItem(BankTabDst, dest, pNewItem);
+        CharacterDatabase.CommitTransaction();
+    }
+    else                                                    // non split
+    {
+        GuildItemPosCountVec gDest;
+        InventoryResult msg = CanStoreItem(BankTabDst, BankTabSlotDst, gDest, pItemSrc->GetCount(), pItemSrc, false);
+        if (msg == EQUIP_ERR_OK)                            // merge to
+        {
+            CharacterDatabase.BeginTransaction();
+
+            if (BankTab != BankTabDst)
+            {
+                LogBankEvent(GUILD_BANK_LOG_MOVE_ITEM, BankTab, pl->GetGUIDLow(), pItemSrc->GetEntry(), pItemSrc->GetCount(), BankTabDst);
+            }
+
+            RemoveItem(BankTab, BankTabSlot);
+            StoreItem(BankTabDst, gDest, pItemSrc);
+            CharacterDatabase.CommitTransaction();
+        }
+        else                                                // swap
+        {
+            gDest.clear();
+            msg = CanStoreItem(BankTabDst, BankTabSlotDst, gDest, pItemSrc->GetCount(), pItemSrc, true);
+            if (msg != EQUIP_ERR_OK)
+            {
+                pl->SendEquipError(msg, pItemSrc, NULL);
+                return;
+            }
+
+            GuildItemPosCountVec gSrc;
+            msg = CanStoreItem(BankTab, BankTabSlot, gSrc, pItemDst->GetCount(), pItemDst, true);
+            if (msg != EQUIP_ERR_OK)
+            {
+                pl->SendEquipError(msg, pItemDst, NULL);
+                return;
+            }
+
+            if (BankTab != BankTabDst)
+            {
+                // check source pos rights (item swapped to src)
+                if (!IsMemberHaveRights(pl->GetGUIDLow(), BankTab, GUILD_BANK_RIGHT_DEPOSIT_ITEM))
+                {
+                    return;
+                }
+
+                // check dest pos rights (item swapped to src)
+                uint32 remRightDst = GetMemberSlotWithdrawRem(pl->GetGUIDLow(), BankTabDst);
+                if (remRightDst <= 0)
+                {
+                    return;
+                }
+            }
+
+            CharacterDatabase.BeginTransaction();
+
+            if (BankTab != BankTabDst)
+            {
+                LogBankEvent(GUILD_BANK_LOG_MOVE_ITEM, BankTab,    pl->GetGUIDLow(), pItemSrc->GetEntry(), pItemSrc->GetCount(), BankTabDst);
+                LogBankEvent(GUILD_BANK_LOG_MOVE_ITEM, BankTabDst, pl->GetGUIDLow(), pItemDst->GetEntry(), pItemDst->GetCount(), BankTab);
+            }
+
+            RemoveItem(BankTab, BankTabSlot);
+            RemoveItem(BankTabDst, BankTabSlotDst);
+            StoreItem(BankTab, gSrc, pItemDst);
+            StoreItem(BankTabDst, gDest, pItemSrc);
+            CharacterDatabase.CommitTransaction();
+        }
+    }
+    DisplayGuildBankContentUpdate(BankTab, BankTabSlot, BankTab == BankTabDst ? BankTabSlotDst : -1);
+    if (BankTab != BankTabDst)
+    {
+        DisplayGuildBankContentUpdate(BankTabDst, BankTabSlotDst);
+    }
+}
+
+void Guild::MoveFromBankToChar(Player* pl, uint8 BankTab, uint8 BankTabSlot, uint8 PlayerBag, uint8 PlayerSlot, uint32 SplitedAmount)
+{
+    Item* pItemBank = GetItem(BankTab, BankTabSlot);
+    Item* pItemChar = pl->GetItemByPos(PlayerBag, PlayerSlot);
+
+    if (!pItemBank)                                     // Problem to get bank item
+    {
+        return;
+    }
+
+    if (SplitedAmount > pItemBank->GetCount())
+    {
+        return;                                         // cheating?
+    }
+    else if (SplitedAmount == pItemBank->GetCount())
+    {
+        SplitedAmount = 0;                              // no split
+    }
+
+    if (SplitedAmount)
+    {
+        // Bank -> Char split to slot (patly move)
+        Item* pNewItem = pItemBank->CloneItem(SplitedAmount);
+        if (!pNewItem)
+        {
+            pl->SendEquipError(EQUIP_ERR_ITEM_NOT_FOUND, pItemBank, NULL);
+            return;
+        }
+
+        ItemPosCountVec dest;
+        InventoryResult msg = pl->CanStoreItem(PlayerBag, PlayerSlot, dest, pNewItem, false);
+        if (msg != EQUIP_ERR_OK)
+        {
+            pl->SendEquipError(msg, pNewItem, NULL);
+            delete pNewItem;
+            return;
+        }
+
+        // check source pos rights (item moved to inventory)
+        uint32 remRight = GetMemberSlotWithdrawRem(pl->GetGUIDLow(), BankTab);
+        if (remRight <= 0)
+        {
+            delete pNewItem;
+            return;
+        }
+
+        CharacterDatabase.BeginTransaction();
+        LogBankEvent(GUILD_BANK_LOG_WITHDRAW_ITEM, BankTab, pl->GetGUIDLow(), pItemBank->GetEntry(), SplitedAmount);
+
+        pItemBank->SetCount(pItemBank->GetCount() - SplitedAmount);
+        pItemBank->FSetState(ITEM_CHANGED);
+        pItemBank->SaveToDB();                              // not in inventory and can be save standalone
+        pl->MoveItemToInventory(dest, pNewItem, true);
+        pl->SaveInventoryAndGoldToDB();
+
+        MemberItemWithdraw(BankTab, pl->GetGUIDLow());
+        CharacterDatabase.CommitTransaction();
+    }
+    else                                                    // Bank -> Char swap with slot (move)
+    {
+        ItemPosCountVec dest;
+        InventoryResult msg = pl->CanStoreItem(PlayerBag, PlayerSlot, dest, pItemBank, false);
+        if (msg == EQUIP_ERR_OK)                            // merge case
+        {
+            // check source pos rights (item moved to inventory)
+            uint32 remRight = GetMemberSlotWithdrawRem(pl->GetGUIDLow(), BankTab);
+            if (remRight <= 0)
+            {
+                return;
+            }
+
+            CharacterDatabase.BeginTransaction();
+            LogBankEvent(GUILD_BANK_LOG_WITHDRAW_ITEM, BankTab, pl->GetGUIDLow(), pItemBank->GetEntry(), pItemBank->GetCount());
+
+            RemoveItem(BankTab, BankTabSlot);
+            pl->MoveItemToInventory(dest, pItemBank, true);
+            pl->SaveInventoryAndGoldToDB();
+
+            MemberItemWithdraw(BankTab, pl->GetGUIDLow());
+            CharacterDatabase.CommitTransaction();
+        }
+        else                                                // Bank <-> Char swap items
+        {
+            // check source pos rights (item swapped to bank)
+            if (!IsMemberHaveRights(pl->GetGUIDLow(), BankTab, GUILD_BANK_RIGHT_DEPOSIT_ITEM))
+            {
+                return;
+            }
+
+            if (pItemChar)
+            {
+                if (!pItemChar->CanBeTraded())
+                {
+                    pl->SendEquipError(EQUIP_ERR_ITEMS_CANT_BE_SWAPPED, pItemChar, NULL);
+                    return;
+                }
+            }
+
+            ItemPosCountVec iDest;
+            msg = pl->CanStoreItem(PlayerBag, PlayerSlot, iDest, pItemBank, true);
+            if (msg != EQUIP_ERR_OK)
+            {
+                pl->SendEquipError(msg, pItemBank, NULL);
+                return;
+            }
+
+            GuildItemPosCountVec gDest;
+            if (pItemChar)
+            {
+                msg = CanStoreItem(BankTab, BankTabSlot, gDest, pItemChar->GetCount(), pItemChar, true);
+                if (msg != EQUIP_ERR_OK)
+                {
+                    pl->SendEquipError(msg, pItemChar, NULL);
+                    return;
+                }
+            }
+
+            // check source pos rights (item moved to inventory)
+            uint32 remRight = GetMemberSlotWithdrawRem(pl->GetGUIDLow(), BankTab);
+            if (remRight <= 0)
+            {
+                return;
+            }
+
+            if (pItemChar)
+            {
+                // logging item move to bank
+                if (pl->GetSession()->GetSecurity() > SEC_PLAYER && sWorld.getConfig(CONFIG_BOOL_GM_LOG_TRADE))
+                {
+                    sLog.outCommand(pl->GetSession()->GetAccountId(), "GM %s (Account: %u) deposit item: %s (Entry: %d Count: %u) to guild bank (Guild ID: %u )",
+                                    pl->GetName(), pl->GetSession()->GetAccountId(),
+                                    pItemChar->GetProto()->Name1, pItemChar->GetEntry(), pItemChar->GetCount(),
+                                    m_Id);
+                }
+            }
+
+            CharacterDatabase.BeginTransaction();
+            LogBankEvent(GUILD_BANK_LOG_WITHDRAW_ITEM, BankTab, pl->GetGUIDLow(), pItemBank->GetEntry(), pItemBank->GetCount());
+            if (pItemChar)
+            {
+                LogBankEvent(GUILD_BANK_LOG_DEPOSIT_ITEM, BankTab, pl->GetGUIDLow(), pItemChar->GetEntry(), pItemChar->GetCount());
+            }
+
+            RemoveItem(BankTab, BankTabSlot);
+            if (pItemChar)
+            {
+                pl->MoveItemFromInventory(PlayerBag, PlayerSlot, true);
+                pItemChar->DeleteFromInventoryDB();
+                StoreItem(BankTab, gDest, pItemChar);
+            }
+
+            pl->MoveItemToInventory(iDest, pItemBank, true);
+            pl->SaveInventoryAndGoldToDB();
+
+            MemberItemWithdraw(BankTab, pl->GetGUIDLow());
+            CharacterDatabase.CommitTransaction();
+        }
+    }
+    DisplayGuildBankContentUpdate(BankTab, BankTabSlot);
+}
+
+void Guild::MoveFromCharToBank(Player* pl, uint8 PlayerBag, uint8 PlayerSlot, uint8 BankTab, uint8 BankTabSlot, uint32 SplitedAmount)
+{
+    Item* pItemBank = GetItem(BankTab, BankTabSlot);
+    Item* pItemChar = pl->GetItemByPos(PlayerBag, PlayerSlot);
+
+    if (!pItemChar)                                         // Problem to get item from player
+    {
+        return;
+    }
+
+    if (!pItemChar->CanBeTraded())
+    {
+        pl->SendEquipError(EQUIP_ERR_ITEMS_CANT_BE_SWAPPED, pItemChar, NULL);
+        return;
+    }
+
+    // check source pos rights (item moved to bank)
+    if (!IsMemberHaveRights(pl->GetGUIDLow(), BankTab, GUILD_BANK_RIGHT_DEPOSIT_ITEM))
+    {
+        return;
+    }
+
+    if (SplitedAmount > pItemChar->GetCount())
+    {
+        return;                                             // cheating?
+    }
+    else if (SplitedAmount == pItemChar->GetCount())
+    {
+        SplitedAmount = 0;                                  // no split
+    }
+
+    if (SplitedAmount)
+    {
+        // Char -> Bank split to empty or non-empty slot (partly move)
+        GuildItemPosCountVec dest;
+        InventoryResult msg = CanStoreItem(BankTab, BankTabSlot, dest, SplitedAmount, pItemChar, false);
+        if (msg != EQUIP_ERR_OK)
+        {
+            pl->SendEquipError(msg, pItemChar, NULL);
+            return;
+        }
+
+        Item* pNewItem = pItemChar->CloneItem(SplitedAmount);
+        if (!pNewItem)
+        {
+            pl->SendEquipError(EQUIP_ERR_ITEM_NOT_FOUND, pItemChar, NULL);
+            return;
+        }
+
+        // logging item move to bank (before items merge
+        if (pl->GetSession()->GetSecurity() > SEC_PLAYER && sWorld.getConfig(CONFIG_BOOL_GM_LOG_TRADE))
+        {
+            sLog.outCommand(pl->GetSession()->GetAccountId(), "GM %s (Account: %u) deposit item: %s (Entry: %d Count: %u) to guild bank (Guild ID: %u )",
+                            pl->GetName(), pl->GetSession()->GetAccountId(),
+                            pItemChar->GetProto()->Name1, pItemChar->GetEntry(), SplitedAmount, m_Id);
+        }
+
+        CharacterDatabase.BeginTransaction();
+        LogBankEvent(GUILD_BANK_LOG_DEPOSIT_ITEM, BankTab, pl->GetGUIDLow(), pItemChar->GetEntry(), SplitedAmount);
+
+        pl->ItemRemovedQuestCheck(pItemChar->GetEntry(), SplitedAmount);
+        pItemChar->SetCount(pItemChar->GetCount() - SplitedAmount);
+        pItemChar->SetState(ITEM_CHANGED);
+        pl->SaveInventoryAndGoldToDB();
+        StoreItem(BankTab, dest, pNewItem);
+        CharacterDatabase.CommitTransaction();
+
+        DisplayGuildBankContentUpdate(BankTab, dest);
+    }
+    else                                                    // Char -> Bank swap with empty or non-empty (move)
+    {
+        GuildItemPosCountVec dest;
+        InventoryResult msg = CanStoreItem(BankTab, BankTabSlot, dest, pItemChar->GetCount(), pItemChar, false);
+        if (msg == EQUIP_ERR_OK)                            // merge
+        {
+            // logging item move to bank
+            if (pl->GetSession()->GetSecurity() > SEC_PLAYER && sWorld.getConfig(CONFIG_BOOL_GM_LOG_TRADE))
+            {
+                sLog.outCommand(pl->GetSession()->GetAccountId(), "GM %s (Account: %u) deposit item: %s (Entry: %d Count: %u) to guild bank (Guild ID: %u )",
+                                pl->GetName(), pl->GetSession()->GetAccountId(),
+                                pItemChar->GetProto()->Name1, pItemChar->GetEntry(), pItemChar->GetCount(),
+                                m_Id);
+            }
+
+            CharacterDatabase.BeginTransaction();
+            LogBankEvent(GUILD_BANK_LOG_DEPOSIT_ITEM, BankTab, pl->GetGUIDLow(), pItemChar->GetEntry(), pItemChar->GetCount());
+
+            pl->MoveItemFromInventory(PlayerBag, PlayerSlot, true);
+            pItemChar->DeleteFromInventoryDB();
+
+            StoreItem(BankTab, dest, pItemChar);
+            pl->SaveInventoryAndGoldToDB();
+            CharacterDatabase.CommitTransaction();
+
+            DisplayGuildBankContentUpdate(BankTab, dest);
+        }
+        else                                                // Char <-> Bank swap items (posible NULL bank item)
+        {
+            ItemPosCountVec iDest;
+            if (pItemBank)
+            {
+                msg = pl->CanStoreItem(PlayerBag, PlayerSlot, iDest, pItemBank, true);
+                if (msg != EQUIP_ERR_OK)
+                {
+                    pl->SendEquipError(msg, pItemBank, NULL);
+                    return;
+                }
+            }
+
+            GuildItemPosCountVec gDest;
+            msg = CanStoreItem(BankTab, BankTabSlot, gDest, pItemChar->GetCount(), pItemChar, true);
+            if (msg != EQUIP_ERR_OK)
+            {
+                pl->SendEquipError(msg, pItemChar, NULL);
+                return;
+            }
+
+            if (pItemBank)
+            {
+                // check bank pos rights (item swapped with inventory)
+                uint32 remRight = GetMemberSlotWithdrawRem(pl->GetGUIDLow(), BankTab);
+                if (remRight <= 0)
+                {
+                    return;
+                }
+            }
+
+            // logging item move to bank
+            if (pl->GetSession()->GetSecurity() > SEC_PLAYER && sWorld.getConfig(CONFIG_BOOL_GM_LOG_TRADE))
+            {
+                sLog.outCommand(pl->GetSession()->GetAccountId(), "GM %s (Account: %u) deposit item: %s (Entry: %d Count: %u) to guild bank (Guild ID: %u )",
+                                pl->GetName(), pl->GetSession()->GetAccountId(),
+                                pItemChar->GetProto()->Name1, pItemChar->GetEntry(), pItemChar->GetCount(),
+                                m_Id);
+            }
+
+            CharacterDatabase.BeginTransaction();
+            if (pItemBank)
+            {
+                LogBankEvent(GUILD_BANK_LOG_WITHDRAW_ITEM, BankTab, pl->GetGUIDLow(), pItemBank->GetEntry(), pItemBank->GetCount());
+            }
+            LogBankEvent(GUILD_BANK_LOG_DEPOSIT_ITEM, BankTab, pl->GetGUIDLow(), pItemChar->GetEntry(), pItemChar->GetCount());
+
+            pl->MoveItemFromInventory(PlayerBag, PlayerSlot, true);
+            pItemChar->DeleteFromInventoryDB();
+            if (pItemBank)
+            {
+                RemoveItem(BankTab, BankTabSlot);
+            }
+
+            StoreItem(BankTab, gDest, pItemChar);
+            if (pItemBank)
+            {
+                pl->MoveItemToInventory(iDest, pItemBank, true);
+            }
+            pl->SaveInventoryAndGoldToDB();
+            if (pItemBank)
+            {
+                MemberItemWithdraw(BankTab, pl->GetGUIDLow());
+            }
+            CharacterDatabase.CommitTransaction();
+
+            DisplayGuildBankContentUpdate(BankTab, gDest);
+        }
+    }
+}
